moved the event system and wallet manager to the common folder. Moved some functions into some util files. Reorganized the files.
