+++ trading-common/src/data.rs
