+++ trading-common/src/lib.rs
@@ -1,14 +1,27 @@
 pub mod constants;
+pub mod data;
 pub mod database;
 pub mod error;
 pub mod models;
 pub mod pumpdotfun;
 pub mod raydium;
-pub mod utils;
+pub mod utils {
+    pub mod copy_trade;
+    pub mod transaction;
+}
+pub mod wallet {
+    pub mod server_wallet_manager;
+}
+
+pub mod events {
+    pub mod event_system;
+}
 
 pub use constants::*;
 pub use database::SupabaseClient;
+pub use events::*;
 pub use models::{ClientTxInfo, CopyTradeSettings, TrackedWallet, TransactionLog, TransactionType};
 pub use pumpdotfun::*;
-
 pub use raydium::*;
+pub use utils::*;
+pub use wallet::*;
